Refactored to remove *most* slider logic out
